Add ViewModels and domain logic
