+++ app/src/main/java/org/ole/planet/myplanet/model/SyncViewModel.kt
@@ -1,0 +1,43 @@
+package org.ole.planet.myplanet.model
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import org.ole.planet.myplanet.callback.SyncListener
+import org.ole.planet.myplanet.service.SyncManager
+
+@HiltViewModel
+class SyncViewModel @Inject constructor(
+    private val syncManager: SyncManager
+) : ViewModel() {
+
+    private val _syncState = MutableStateFlow<SyncState>(SyncState.Idle)
+    val syncState: StateFlow<SyncState> = _syncState.asStateFlow()
+
+    fun startSync(type: String = "full", tables: List<String>? = null) {
+        syncManager.start(object : SyncListener {
+            override fun onSyncStarted() {
+                _syncState.value = SyncState.Syncing
+            }
+
+            override fun onSyncComplete() {
+                _syncState.value = SyncState.Success
+            }
+
+            override fun onSyncFailed(msg: String?) {
+                _syncState.value = SyncState.Failed(msg)
+            }
+        }, type, tables)
+    }
+
+    sealed class SyncState {
+        object Idle : SyncState()
+        object Syncing : SyncState()
+        object Success : SyncState()
+        data class Failed(val message: String?) : SyncState()
+    }
+}
+
