+++ app/src/main/java/org/ole/planet/myplanet/ui/dashboard/DashboardActivity.kt
@@ -90,12 +90,12 @@ import org.ole.planet.myplanet.utilities.Constants
 import org.ole.planet.myplanet.utilities.Constants.showBetaFeature
 import org.ole.planet.myplanet.utilities.DialogUtils.guestDialog
 import org.ole.planet.myplanet.utilities.EdgeToEdgeUtil
-import org.ole.planet.myplanet.utilities.FileUtils.totalAvailableMemoryRatio
 import org.ole.planet.myplanet.utilities.KeyboardUtils.setupUI
 import org.ole.planet.myplanet.utilities.LocaleHelper
 import org.ole.planet.myplanet.utilities.NotificationUtil
-import org.ole.planet.myplanet.utilities.TimeUtils.formatDate
 import org.ole.planet.myplanet.utilities.Utilities.toast
+import org.ole.planet.myplanet.domain.NotificationCreator
+import org.ole.planet.myplanet.domain.ChallengeEvaluator
 
 @AndroidEntryPoint  
 class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, NavigationBarView.OnItemSelectedListener, NotificationListener {
@@ -116,6 +116,8 @@ class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, N
     private val dashboardViewModel: DashboardViewModel by viewModels()
     private lateinit var challengeHelper: ChallengeHelper
     private lateinit var notificationManager: NotificationUtil.NotificationManager
+    private lateinit var notificationCreator: NotificationCreator
+    private lateinit var challengeEvaluator: ChallengeEvaluator
     private var notificationsShownThisSession = false
 
     private interface RealmListener {
@@ -132,6 +134,8 @@ class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, N
         initViews()
         updateAppTitle()
         notificationManager = NotificationUtil.getInstance(this)
+        notificationCreator = NotificationCreator(dashboardViewModel, notificationManager)
+        challengeEvaluator = ChallengeEvaluator(mRealm, user, settings, editor)
         if (handleGuestAccess()) return
         setupNavigation()
         handleInitialFragment()
@@ -141,7 +145,7 @@ class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, N
         setupRealmListeners()
         checkIfShouldShowNotifications()
         addBackPressCallback()
-        challengeHelper = ChallengeHelper(this, mRealm, user, settings, editor, dashboardViewModel)
+        challengeHelper = ChallengeHelper(this, mRealm, user, settings, editor, challengeEvaluator, dashboardViewModel)
         challengeHelper.evaluateChallengeDialog()
         handleNotificationIntent(intent)
     }
@@ -516,7 +520,7 @@ class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, N
             try {
                 databaseService.realmInstance.use { backgroundRealm ->
                     backgroundRealm.executeTransaction { realm ->
-                        val createdNotifications = createNotifications(realm, userId)
+                        val createdNotifications = notificationCreator.createNotifications(realm, userId)
                         newNotifications.addAll(createdNotifications)
                     }
 
@@ -543,131 +547,6 @@ class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, N
         }
     }
 
-    private fun createNotifications(realm: Realm, userId: String?): List<NotificationUtil.NotificationConfig> {
-        val newNotifications = mutableListOf<NotificationUtil.NotificationConfig>()
-
-        dashboardViewModel.updateResourceNotification(realm, userId)
-
-        newNotifications.addAll(createSurveyNotifications(realm, userId))
-        newNotifications.addAll(createTaskNotifications(realm, userId))
-        newNotifications.addAll(createStorageNotification(realm, userId))
-        newNotifications.addAll(createJoinRequestNotifications(realm, userId))
-
-        return newNotifications
-    }
-
-    private fun createSurveyNotifications(realm: Realm, userId: String?): List<NotificationUtil.NotificationConfig> {
-        val newNotifications = mutableListOf<NotificationUtil.NotificationConfig>()
-        val pendingSurveys = dashboardViewModel.getPendingSurveys(realm, userId)
-        val surveyTitles = dashboardViewModel.getSurveyTitlesFromSubmissions(realm, pendingSurveys)
-
-        surveyTitles.forEach { title ->
-            val notificationKey = "survey-$title"
-
-            if (!notificationManager.hasNotificationBeenShown(notificationKey)) {
-                dashboardViewModel.createNotificationIfNotExists(realm, "survey", title, title, userId)
-
-                val config = notificationManager.createSurveyNotification(title, title)
-                newNotifications.add(config)
-            }
-        }
-        return newNotifications
-    }
-
-    private fun createTaskNotifications(realm: Realm, userId: String?): List<NotificationUtil.NotificationConfig> {
-        val newNotifications = mutableListOf<NotificationUtil.NotificationConfig>()
-        val tasks = realm.where(RealmTeamTask::class.java)
-            .notEqualTo("status", "archived")
-            .equalTo("completed", false)
-            .equalTo("assignee", userId)
-            .findAll()
-
-        tasks.forEach { task ->
-            val notificationKey = "task-${task.id}"
-
-            if (!notificationManager.hasNotificationBeenShown(notificationKey)) {
-                dashboardViewModel.createNotificationIfNotExists(
-                    realm,
-                    "task",
-                    "${task.title} ${formatDate(task.deadline)}",
-                    task.id,
-                    userId
-                )
-
-                val config = notificationManager.createTaskNotification(
-                    task.id ?: "task",
-                    task.title ?: "New Task",
-                    formatDate(task.deadline)
-                )
-                newNotifications.add(config)
-            }
-        }
-        return newNotifications
-    }
-
-    private fun createStorageNotification(realm: Realm, userId: String?): List<NotificationUtil.NotificationConfig> {
-        val newNotifications = mutableListOf<NotificationUtil.NotificationConfig>()
-        val storageRatio = totalAvailableMemoryRatio
-        val notificationKey = "storage-critical"
-
-        if (storageRatio > 85 && !notificationManager.hasNotificationBeenShown(notificationKey)) {
-            dashboardViewModel.createNotificationIfNotExists(realm, "storage", "$storageRatio%", "storage", userId)
-
-            val config = notificationManager.createStorageWarningNotification(storageRatio.toInt())
-            newNotifications.add(config)
-        }
-        return newNotifications
-    }
-
-    private fun createJoinRequestNotifications(realm: Realm, userId: String?): List<NotificationUtil.NotificationConfig> {
-        val newNotifications = mutableListOf<NotificationUtil.NotificationConfig>()
-        val teamLeaderMemberships = realm.where(RealmMyTeam::class.java)
-            .equalTo("userId", userId)
-            .equalTo("docType", "membership")
-            .equalTo("isLeader", true)
-            .findAll()
-
-        teamLeaderMemberships.forEach { leadership ->
-            val pendingJoinRequests = realm.where(RealmMyTeam::class.java)
-                .equalTo("teamId", leadership.teamId)
-                .equalTo("docType", "request")
-                .findAll()
-
-            pendingJoinRequests.forEach { joinRequest ->
-                val notificationKey = "join_request-${joinRequest._id}"
-
-                if (!notificationManager.hasNotificationBeenShown(notificationKey)) {
-                    val team = realm.where(RealmMyTeam::class.java)
-                        .equalTo("_id", leadership.teamId)
-                        .findFirst()
-
-                    val requester = realm.where(RealmUserModel::class.java)
-                        .equalTo("id", joinRequest.userId)
-                        .findFirst()
-
-                    val requesterName = requester?.name ?: "Unknown User"
-                    val teamName = team?.name ?: "Unknown Team"
-                    val message = "$requesterName has requested to join $teamName"
-
-                    dashboardViewModel.createNotificationIfNotExists(
-                        realm,
-                        "join_request",
-                        message,
-                        joinRequest._id,
-                        userId
-                    )
-
-                    val config = notificationManager.createJoinRequestNotification(
-                        joinRequest._id!!,
-                        requesterName,
-                        teamName
-                    )
-                    newNotifications.add(config)
-                }
-            }
-        }
-        return newNotifications
-    }
     private fun openNotificationsList(userId: String) {
         val fragment = NotificationsFragment().apply {
             arguments = Bundle().apply {
